requirements.txt
albumentations==1.3.0
django_recaptcha==3.0.0
matplotlib==3.6.3
numpy==1.23.5
Pillow==9.4.0
pytorch_lightning==1.9.4
textdistance==4.5.0
torch==1.11.0+cu113
torch_summary==1.4.5
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.1
~pencv_python==4.5.3.56
